requirements.txt
flask
torch
diffusers
transformers
accelerate
safetensors
python-dotenv
